Simplify xmlHttpRequest bridge
